added tests for CommentDao

+++ books-maslov/src/main/java/com/maslov/booksmaslov/repository/CommentDao.java
@@ -2,11 +2,10 @@ package com.maslov.booksmaslov.repository;
 
 import com.maslov.booksmaslov.domain.Comment;
 
-import java.util.List;
-
 public interface CommentDao {
 
-    Comment getCommentById(int id);
+    Comment getCommentById(long id);
+
     Comment createComment(String comment);
 
     void updateComment(Comment comment);
